+++ GWToolboxdll/GWToolbox/GuiUtils.h
@@ -15,7 +15,7 @@ namespace GuiUtils {
     bool FontsLoaded();
 	ImFont* GetFont(FontSize size);
 
-	int GetPartyHealthbarHeight();
+    size_t GetPartyHealthbarHeight();
 
 	std::string ToLower(std::string s);
 	std::wstring ToLower(std::wstring s);
